requirements.txt
flask
cryptography
kyber-py
flask-talisman 
python-dotenv